backend/requirements.txt
# Seedy Backend
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*
httpx==0.28.*

# RAG
qdrant-client==1.13.*

# Reranker
sentence-transformers==3.*
torch>=2.0

# Ingestion
python-docx==1.*
PyPDF2==3.*

# Utils
python-dotenv==1.*

# MQTT (para publicar eventos vision)
paho-mqtt>=2.0

# Vision (thumbnails de aves)
Pillow>=10.0

# YOLO (detección local de aves)
ultralytics>=8.3
opencv-python-headless>=4.9
